Fix migration discovery issue
When devices are migrated, we must pull the serial number from the
device data rather than the device_network_id (which is what devices
joined via the driver use) because the DTH backed devices use MAC
addresses for the dni, which are not available to edge driver devices.

This fixes a bug with discovery on device init where when multiple
devices are discovered at the same time, the last discovered device IP
is used for all the websocket connections. This situation happens
exclusively during migration