Fix max codes calculation to allow for COTA cred to be set
Previously calculating the max credential index available for pin codes
we used the formula `NumberOfPINUsersSupported *
NumberOfCredentialsSupportedPerUser`, but this calcuation was misguided
and results in an index out of range for the device. The max credential
index available on a lock is actually bounded by
`NumberOfPINUsersSupported`. This is not super clear in the spec, but
has been confirmed by some of the lock companies, and other CSA members.