Set COTA credential at lowest available index
Upon receiving RequirePINforRemoteOperation we kick of the process of
setting the COTA credential. SetCredential is issued at index 1, if it
fails, the response contains the lowest available index and we set the
credential there. This is an improvement from clearing and setting in
the highest available credential index because it doesn't clear
credentials set by the user.