BUG-6683 Simplify CT100 driver
A lot of behavior was ported over from the DTH which, I think, in hindsight, could have been foregone. I've attempted to pare this driver back to the z-wave standard except for some polling that seems as though it might have been necessary. This also includes a new profile with a range defined to match the set point bounds in the DTH.

device polls humidity and temperature using multi-channel encapsulation

change temp scale detection based on device limits