fix: Perform explicit MAC lookup at discovery
Some bridges won't use their MAC-based Bridge ID as the hostname when responding
to mDNS queries. We were relying on that functionality in our discovery loop so
we fix that by querying the bridge config info, which is an endpoint that does
not require auth and that has the MAC address as an explicitly reported field.

Fixes: CHAD-10160