fix: Dimming ranges and dimming behavior.
Hue has per-product dimming values, and it would be nice to accomodate those. But
there are two problems. The first is that Hue allows for fractional values,
the second is that the min-dimming is often < 1. The DTH simply treated all hue
devices as having a minimum dimming value of 1. So we do that here to maintain
behavior parity with the DTH.

We also add ranges for dimming to the profile yaml.

Fixes: CHAD-10099