Wemo MAC differences on migrated devices
The MAC printed on the device and advertised in device metadata is used for the
access point that the wemo device sets up for onboarding, and the MAC that is
used once on the homes network is one more than what is advertised in metadata.
The DTH got the actual MAC because its discovery pulls this info from the network;
the driver queries the device for this info because it is not available on the socket
that we have in drivers.

An alternative approach would be to "calculate" the actual mac address
from what is reported from the device, but there is no guarantee that
all wemo devices will actually being doing this MAC change.