First attempt at subscription server
Managing the socket and ctrl channel on a single task with select failed
because of a hub firmware bug where select never returns a socket ready to
accept a connection. This resulted in the server being managed
separately from the control channel. There were still issues though
since setting a timeout on a tcp server socket resulted in a driver
crash, which is also a hub firmware bug. This means it is unavoidable to
not leak tasks that are waiting on accept; not the biggest deal.