fix: Eliminate several spurious refreshes
The logic revised in this commit was causing a lot of spurious
events to be emitted, so we revise some things.

First is that we only refresh after init on fresh join, not general init.

Second is that we don't do a forced refresh whenever the SSE stream connects.
This is fine for fresh join because we do a forced refresh anyway, and it's
fine for reconnects because the SSE implementation will request any messages
that it has missed in the interrim whenever it performs a reconnect.