fix: Responsive attr. emit after handling commands
Previously we relied on the eventual update from the SSE stream
to emit capability attribute updates that result from command
handling. This is subject to latency spikes and responsiveness
jitter due to encapsulation and rate limiting on the Hue bridge
itself that it uses to conserve its own resources.

This change adopts a pattern found in other drivers where the endpoint
provides an ack which we use as a signal that we can emit and update
that attribute update that matches the args to command we just sent.
We then cache that attribute update and we check against that for similarity
before performing any additional attribute event emissions for that capability..