Include "replace" action to playPreset Command
In the absence of the replace parameter shown here, if a user selects a Favorite that is a playlist (as opposed to a station), the playPreset command *appends* the Favorite selection to the end of any existing queue rather than replacing it as expected. This not only causes the queue to continue growing in length until manually cleared via the Sonos app or Desktop interface, but if the user has "shuffle" enabled, the intended list of songs are shuffled into the larger queue and do not play directly in response to the request.    

Note: The replace command has no effect on stations, only playlists.

P.S. The option to toggle "shuffle" on and off should also be present.  

{
  "favoriteId" : "223",
  "playOnCompletion" : true,
  "action" : "replace"
  },
  "playModes" : {
    "shuffle" : false
  }
}