fix(zwave-lock): Update a faulty setCode test
There was a bug in the z-wave defaults that didn't get
caught by this test but should have. The bug: When we
were emitting code events, we were always using the fallback
form of `Code N` even if a user supplied lock code name
was set.

The test as written uses the fallback code name as the user
supplied code name, meaning that the test didn't catch the
edge case when the user supplied a name that wasn't the same
as the default.