fix: Streamline device initialization process
The initialization routine opted for a pretty heavy-weight eventual
consistency model with a bunch of repeated/duplicated checks on
speakers as they onboarded. It was also doing an extra round
or two of re-scanning.

This tightens up some of the checks to prevent doing a bunch of
un-needed work and introducing a bunch of un-needed network traffic.