fix(sonos): Don't onboard non-primary bonded devices.
If a device is part of a bonded set it would not be onboarded by
the DTH. But there was a bug in the edge driver that would bring
these devices over after migration and join them fresh during
discovery when using the Edge Driver.

This fix gets us back to parity with the DTH by excluding bonded
devices that are not the primary device.