BUG-7616 Wait for response before sending follow-up polling with setpoint gets
Users have reported issues controlling the setpoints of CT100 thermostats. Inspecting logs revealed that the device was not responding to setpoint gets. This was presumably due to issues with the device being overloaded by the number of z-wave messages in a short timeframe. Timers were implemented to attempt to mitigate this, but on busy networks, message queueing means that delays between commands may not be preserved when they're sent by the radio. This approach waits until we have received a response and then sends any subsequent follow-up polls.