Bugfix: Bose devices sometimes marked offline "forever"
The driver attempted to reconnect a devices websocket at most 10 times,
before stopping. This meant that if the device did not come back online
within 20 minutes, the device would be offline forever, or at least
until the driver is restarted or the device is refreshed while it is
reachable on the network.