Refactor migration to avoid over-spawning tasks
We introduced some accidentally quadratic behavior that was leading
to a massive amount of cosock tasks getting spawned.

We had to refactor the data flow a little bit in order to achieve this,
so the behavior may appear to have a changed a bit more than preferred.
However, the savings on CPU/Memory/network overhead are appreciable.