fix: Allow SSDP "rediscovery" to run indefinitely
This moves the SSDP Rediscovery process (re-scanning during added or
init) on to a cosock task, and replaces the upper bound on attempts
with retry-until-success logic.

We have to move this on to a task if we want to retry until we succeed
otherwise we'd be blocking the device's thread in the lifecycle handlers.