Switch generic locks that support lock codes
Because our fingerprinting system cannot distinguish between zigbee locks that do and do not support lock codes, users with non-WWST-certified devices that do support lock codes are unable to use that functionality on our platform. Previously they could easily switch the DTH to enable this, but with drivers this is not possible without a custom driver installed.

This logic attempts to determine whether a lock that was fingerprinted without user codes actually supports user codes or not, and switches the profile to one which does. It will also attempt to recover and lock codes carried over from migration.