Fix luacheck violations from vendored Cosock file.
The file in question tries to capture the environment's
canonical table pack/unpack operations, which use a global
function in older luas but a function on the `table` namespace
in newer Luas.

Since those globals don't exist in 5.3, it's a violation, and
we don't want that invalidating our CI observability.