fix: Periodic SSDP scan to detect IP changes
Similarly to Hue, this adds a periodic scan for detecting changes in IP address
at runtime after a device has onboarded.

This refactor isn't as surgical as the Hue one because the Sonos topology is
simpler, so we didn't do a deep gutting of the discovery logic and left it
mostly untouched, so we still actively perform SSDP scans during discovery.

Relaxing this for Sonos is likely best left to a refactor after we move the
SSDP APIs in to hubcore and expose them via the lua libs.

Fixes: CHAD-11472