fix: Check IPV4 field when detecting change
In the specific situation where a Hue topology gets migrated with a stale IP
address, we can get in to a situation where the change detection happens
before the `[joined_bridges]` map is populated, meaning that subsequent scans
won't know that an update needs to happen.

We rectify that by also factoring in the stored IP in the device's fields
when determining if a change has occurred.