fix: Check DNI + parent assigned key for strays
This is a result of a workaround for a pre-release defect that was present
during pre-release testing and has been fixed but we need to continue to
support devices that may have encountered it.

For some devices the service ID will be captured in the DNI instead of the
parent assigned child key and so we should check both, but due to an oversight
we are only looking at the DNI currently.

This could assist in some rare situations where the v1 ID may not be stable, but
I don't have any evidence that this bug is actually impacting any users.