fix: Patched cosock.lua to avoid spurious crash
A race condition bug in Cosock can lead to crashes, and we think we have a good
bead on fixing the issue with these changes. Because these updated lua libs
won't go out for a while, we're investigating keeping them as an in-tree patch
to alleviate the issues in the driver for now.