Do not emit event when color temp outside of capability temp range
Some devices could report a value of 0 mireds, which would result
in a divide by zero error when converting to kelvin. This
introduces a check to make sure that 0 and other mired values
that are outside the range of the colorTemperature capability are
not handled since they are likely not real values.