Add datastore example
This changs adds an example for how to use the datastore. It also
fixes a typo to go with changes in lua_libs. Previously the api would
expect a "partner" argument with get_aqara_secret, however this was
removed in a previous change but the check on the argument was missed so
that command would always fail.