Matter Switch: fix endpoint to component mapping
The matter switch driver was initially written with the assumption
that for multi-endpoint devices the endpoints would be sequentially
numbered starting from 1. However, this is not a requirement in the
matter spec, and devices like composed bridged devices often do
not follow this assumption. To address this, endpoints are now
queried on device add and then mapped to the corresponding
component.