BUG-8651 Aeotec Nano Shutter: add missing event
The stateless curtain power button needs to set its available command values before it will function correctly in the app. This was done in the DTH but was missed in the porting work.

I've added changes so that a refresh of the device should bring an already-installed device back to a good state. These changes can be removed at a later date as future installs will have them from the added handler changes.