Matter Thermostat: fix default endpoint mapping
Our "default" endpoint logic current returns the lowest endpoint.
For matter composed bridged devices, the current "default" endpoint
would be the BRIDGED_NODE_DEVICE_TYPE endpoint because it is the
lowest endpoint for the bridged device. However, that endpoint does
not support the desired clusters, so we want to find the first endpoint
that supports our desired cluster and map that to the corresponding
component.

This should not affect existing behavior because for single endpoint
non-bridged devices, the default endpoint would also support the
given cluster.