matter-button: Fix multi-button devices that don't start at endpoint 1
This fixes two problems with the endpoint/component mapping:
- The Lua table serializer cannot handle numerical keys that don't start
  at 1 (aka an array) because it serializing to JSON and JSON objects
  require keys to be strings.
- Using `#` to check the size of a Lua table only works if the keys start
  at 1

Due to the second issue, the driver wasn't even trying to write the map
to persistance, but even if that was fixed the first issue caused the driver
to crash.

To fix the first issue we switch to using the components as the key which
are strings so there is no issue converting to/from JSON.

To fix the second issue we use another variable to determine if we should
write the map to persistance.

With these changes and the 0.50 hub firmware the Hue multi-button devices
are now working.

https://smartthings.atlassian.net/browse/MTR-453