matter-switch: Fix devices (ex: Hue White) that did not have levelControl
In Matter it is valid for an endpoint that supports a specific device type
to also support additional device types that are subsets of that device type.
The Hue White is an example of a device that does this. Since we didn't
have a fingerprint for an endpoint with both OnOff Light and Dimmable Light
device types the fingerprinting logic happened to be picking the fingerprint
with just the OnOff device type which uses a profile without levelControl.

The fix is to add additional fingerprints for each of the device types that
are a superset of other device type(s).

https://smartthings.atlassian.net/browse/MTR-473