matter-button: Fix warning related to battery with Hue switches/buttons
Hue accessories like the Tap Dial Switch that use the Matter Button driver
do support the PowerSource cluster but they do not implement the optional
BatPercentageRemaining attribute. The `battery` capability requires the
percentage remaining so we're going to remove the `battery` capability from
these devices for now.

We have plans to replace this code with something more dynamic in the future,
plus we want to expose the battery information for these devices somehow.

Note: This change will only apply to newly joined devices.

https://smartthings.atlassian.net/browse/MTR-474