matter-switch: remove unsupported and duplicate fingerprints
The OnOff Switch, Dimmer Switch, and Color Dimmer Switch are
all device types that requiring bindings in order to work
correctly. At the moment we do not support bindings, so we
will remove these fingerprints so the devices will join
as matter-things instead.

Also, removes a duplicate fingerprint for matter dimmable
switch.