fix: Only allow a single discovery handler
There are some situations where many discovery handlers can be started
at the same time, and in some cases this can lead to issues like duped
devices being created because the event create requests happen close
enough together that they skirt through checks for things like DNI
duplication.

The proper fix for this is to disallow this in firmware, but this is
impacting users currently so a stop-gap fix in the LAN Edge Drivers
at a minimum -- where this is more likely to happen -- may not be a bad
idea.