fix: Remove some lookups from persistent store
Subtle edge cases with the persistent store's eventual consistency mean
that it's possible for these stores to get in to a bad state, and some
of the APIs on the hub will hard crash when given stale data. Moving
these stores to be ephemeral could slow down detection of IP address
changes but shouldn't hamper functionality.