matter-button: Missing button press after a long press
Some Switch devices will send a MultiPressComplete event at the end of a
long press sequence and others will not. Normally the driver creates button
capability events on receipt of the MultiPressComplete but in the case of
a long press it creates the "held" event on receipt of the LongPress event.
If it also receives a MultiPressComplete as part of that long press sequence
then it should ignore it so as not to generate an incorrect extra event.

The problem was that if a new button press was started (as determined by the
receipt of a InitialPress) the driver was still ignoring the next MultiPressComplete
event, and so that button press sequence would be ignored.

The fix is to only ignore MultiPressComplete that occurs as part of a long
press sequence.

https://smartthings.atlassian.net/browse/MTR-509