Support switching from matter-thing to switch-binary
The On/Off Light switch device type requires bindings in order to
work correctly because the On/Off cluster on this device type is
client type as specified in the matter spec. Because we do not
support bindings, we have removed the generic fingerprint for this
device type so that it will join as a matter-things so we do not
expose broken funcitonality to users. However, we have found some
devices that do not follow the spec and report an On/Off Light
Switch device type and yet have a server type OnOff cluster.
These out-of-spec devices would technically have worked before
we decided to remove the fingerprints due to lack of binding
support, so these devices will be broken now that they will join
as matter-thing. This adds a workaround that allows us to check
for a special case where a matter-thing supports the OnOff cluster
as a server, and then we switch to the switch-binary profile.

Additionally, this improves our multi-switch profile switching
logic to only count server type On/Off cluster endpoints towards
the switch total for the same reasons as described above.