Fix parsing patterns for SSDP responses
The current way we parse the SSDP response headers can cause failures
in discovering Sonos speakers with unicode characters that require
more than one byte for their representation, such as CJK characters.

The current parsing patterns use the '`%g`' character class, which does
not parse characters that don't fit in a single byte; while Lua itself
is utf-8 compatible and it is safe to use multi-code-point graphemes or
clusters in Lua strings, the actual string library is byte-oriented for
many operations and the pattern used currently tries to match each byte
against the provided class.

'`.-`' works here instead, as it matches all bytes in a non-greedy way
(akin to '`.*?`' in PCRE).