BUG2-659 Zigbee locks: fix out-of-order commands
Commands can arrive from a zigbee lock after an unlock event in any order. This can be a problem if the door lock operation event arrives _after_ the attribute report about the door lock state. Because the door lock attribute update doesn't include any useful metadata, but it arrives first, the second command with the useful metadata is marked as a duplicate and debounced, dropping all of that useful metadata on the floor. This same issue can also happen with Z-Wave, but because of the overloading of the notification defaults, it will have to be fixed via updates to the lua libraries. This fix is also going into the lua libraries, but it's easier to also deploy it quicker here.