Matter Switch: prevent matter bridges from switching to MCD
Some matter bridge devices may incorrectly report additional
endpoints with on/off support, which would then make them switch
to multi-switch profiles. In order to prevent this check for
the "aggregator" device type on init, and prevent a profile
switch to an MCD switch profile if it is present.

Existing bridge devices that have been switched to an MCD profile
will remain as MCD devices.