CHAD-12259 zigbee motion: high battery usage
Users reported higher battery consumption with edge drivers and a disparity was noted between DTH and driver cluster configurations. A forthcoming change in the lua libraries will change the defaults, but this change should significantly improve the situation for most users in the meantime with a simple refresh. Changes the IAS Zone reporting configuration to use the device defaults and changes temperature reporting to a max interval of 10 minutes with a 1 degree celsius reporting threshold.

add change for zigbee contact as well