Add line numbers to failure outputs for non-ci run
This change adds line number output to failure reports from the non-ci
`run_driver_tests.py` script. This script doesn't run in CI/CD but can
be useful for local development since it can take the `--filter` arg,
which lets developers run groups of tests at once with JUnit reports.

Adding the line numbers to the `stdout` report is a QoL improvement
for devs to jump to tests quickly; I've even been able to create a
problem matcher for the output in VS Code now.

The `run_driver_tests_p.py` script that runs in CI/CD is unmodified.