Z-Wave Switch: lazy load sub drivers
Adds the ability to lazy load sub drivers. Sub driver that are
not currently used by the driver will be "lazy loaded", meaning
just a smaller version of the sub driver will be loaded until
a device joins that requires the entire sub driver, at which point
all the sub driver handlers will be loaded. Given the high number
of sub drivers in the zwave-switch driver, this should provide a
good amount of memory savings in cases where many of the sub
drivers go unused.

Note, this will only work with the latest lua libs changes, so the
lazy loading is only activated after an API version check.