fix: Add some hardening to the Hue discovery logic.
The first thing we do is we store the API keys in the datastore right
away instead of waiting for a device record to exist. This hardens us
against the situation where the driver restarts for any reason while
the Hue Bridge device's `added` handler is still executing and is early
enough in its execution that we haven't stored the key on the device
object yet.

We also address an issue where an invalid mDNS response (for example,
an ipv6 response) would `return` from the entire mDNS scan function
instead of moving on to the rest of the responses returned. In practice,
this is not likely to happen, as we don't typically get both an ipv4 and
and ipv6 from a Hue bridge, but I noticed it while doing this refactor
and felt it was worth addressing.

We also address a regression from the automatic add/delete changes,
which can lead to the creation of duplicate device records for bulbs in
some specific circumstances.