Matter Switch: add handling for multi-On/Off Light Switch device types
The On/Off Light Switch type supports the On/Off cluster as client,
however some deviecs still report the On/Off light switch but
implement On/Off cluster as server or both. There was previously
a check to handle this case, but it did not account for
multi-devices, so this expands the check to handle multi-devices
that fall in this category as well.