Hotfix: Revert auto-add/auto-delete on 0.47 hubs
This change puts an API Version guard around the call that performs the
auto-add/auto-delete scan at startup, which is causing a crash on 0.47
TV's due to an issue with looking up non-existent UUID's.

We believe the root of the issue is that the non-existent UUID lookup
is related to trying to look up the Hub UUID as the parent of the bridge,
but because this is a pretty gnarly issue we're going to roll back
while we're confirming our suspicion.