Skip processing bridges as deletion candidates
This removes the API version guard on the auto add/delete scan during
event stream connection and instead skips to the next element in the
iterator if it detects that the current iteration target is a device
record that represents a bridge.

We do this because we *do* want the auto-add/auto-delete routine to
run on older firmware, as the delete function does feature detection
on the lua libs and uses `device:offline()` as a fallback. Plus, auto
add will work on all firmware versions, so this is the better UX all
around.