Fix the repair task for unlinked bridges
Sometimes a data loss event can occur that prevents the Bridge
Application Key from being persisted in the Edge Driver Datastore.

This coroutine broke at some point during a refactor, unsure when, but
this commit restores the functionality that would let a user re-link
their bridge in the event that they're aware that this is what has
happened.

Depends on #1211