Matter Switch: refresh child devices on added event
There is a race condition where sometimes the initial subscription
report will be handled before child devices are done being created.
In this case, the child devices will not recieve the initial value
of the attribute, and will appear as just "connected" or "waiting
for device to update" until they are  manually refreshed or the
state is physically toggled on the device.