Eve Energy: temporarily exclude Eve Outlet from sub driver
Temporarily exclude Eve Outlet from the eve-energy sub driver
while waiting for WWST to complete. This will allow the outlet
device to use the parent driver so that it can utilize the parent
child logic in the parent driver to create a device card for
each outlet instead of just having control over one outlet.
Power support will not be available until the other eve changes
are merged with WWST.