Eve Energy: fix child can_handle and get switch state on add
When a child device is added, request the switch state. This is
done to account for a race condition where a subscription report
arrives before the child device has been created and the child
device misses the initial state.

This also fixes an issue where can_handle was not returning true
for the child device.