Matter Switch: only set_find_child for parent-child devices
Only set the find_child function for parent-child devices. This is
because in cases where a bridge is onboarded with many devices,
the set_find_child function is call which sets the find_child
function. When the bridge is onboarded, it then calls get_devices()
when when the first event is emitted because the device checks for
any child devices the current device might have. This get_devices()
call ends up adding all devices in the driver to the device cache.
Once in the device cache, the `init` event is skipped. However,
there is a race condition where when adding a large number of
bridged devices, some devices can be added to the device cache
before they have initialized, causing them to miss the `init`
event. This happens when a device with a find_child function
emits an event for a specific endpoint, because when it emits
and event for an endpoint, it calls the find_child function
to see if there is a child device that it should emit the event for,
which causes the get_devices() functions to be called as described
above, adding some as yet uninitialized devices to the device
cache incorrectly.