matter-rvc: Modify the mode capability to update normally
This change solves the problem that mode capability is not updated
normally in the plug-in of the app after the Matter RVC device is
onboarded.
If the device transmits the attributes values of multiple clusters at
once, it is not known in which order the driver will call the
CurrentMode and SupportedMode handlers.
This change improves the driver to support this case because the
existing driver did not consider cases that call in the order of
SupportedModes and CurrentMode handlers.

IOTE-3658
IOTE-3768

Signed-off-by: Youngho Yoon <34558998+yhoyoon@users.noreply.github.com>