Refactor discovery for addition of new device types.
This is mostly a structural refactor as opposed to a functional
one, but there are subtle changes.

First, we refactored the discovery module in to a directory, as
we're going to be breaking the file up. The first extraction here is
moving the handler for processing a discovered light in to its own file.

Next, we changed some names of some variables where it was assumed that
we were working with lights; this is purely a comprehension/readability
refactor.

The only functional change made here was a modification to the
`supported_resource_types` mapping, which was originally just a map
of booleans. This is now a map from a resource type to a handler function.
In this case, we import the handler function from the new file mentioned
above.