Refactor the bridge refresh-all functionality.
The current refresh-all functionality assumed that all child devices
were lights and immediately hit that endpoint, then did a bunch of
light-specific work.

This refactor instead hits the API endpoint to get the state for types
that are present on child devices in the driver and then dispatches
to the appropriate handler per child device after all relevant service
states are queried and cached.