Add new utils functions for device type detection.
Previously, because we only supported bridges and lights, there
was a good deal of logic that checked for a device being a light simply
by seeing if it was *not* a bridge.

Now that we need to support more than lights, this won't suffice.

This is further complicated because we cannot rely on the presence of any
information introspected during Discovery to still be around at `added`.

We introduce a new scheme for the parent-assigned child key to solve this, as
we know this will be available on the device record at the time the `added`
handler is invoked.

We also maintain backward compatibility with the light devices that were either
migrated or joined before this refactor took place.