Support for compound devices and "grandchildren"
This refactor furthers the generalizations done previously for devices
that provide multiple services to allow for not only mapping of services
to capabilities or components, but also multiple device records.

This allows for us to support devices such as the Dymera up/down light.

Other changes:

- I found myself frequently introducing bugs related to using the Hue Bridge ID
(MAC Address) as a Device ID, and the Bridge Device ID as a Hue Bridge ID. We renamed
all instances of `bridge_id` referencing the MAC to `bridge_network_id` to try and
alleviate the confusion.

- Button support is further generalized to support two-button models, which covers the
In-Wall Switch Module in the 2-gang configuration.