zigbee-contact: Fix garage door handling for SmartSense Multi
The original SmartSense Multi has a different z-axis than the newer sensors.
When installed on the garage door it is parallel to the ground when the door
is closed and perpendicular to the ground when the door is open. For this
reason it can't share the garage door handling code that the newer sensors
use. This change implements the necessary custom handling and adds a unit
test.

https://smartthings.atlassian.net/browse/CHAD-13098