Matter-Window-Covering: Fix refresh bug
When the device card for a window covering device is refreshed, the
shade level displays as 0%. This is due to the CurrentLevel attribute,
which is not a required clusted for the window covering device type,
overwriting the windowShadeLevel capability after is it set by
CurrentPositionLiftPercent100ths. This PR removes the CurrentLevel
attribute from the matter-window-covering driver.