Feat: Busted tests for Hue driver
We don't run Busted tests in CI, but the Hue driver is now very
testable after the recent refactor. The ability to write integration
tests for LAN Drivers is still forthcoming, but we're also going to be
accumulating unit tests for Hue going forward as well due to the fact
that it's huge + complex but is now much more testable than it used
to be.

We keep the `spec` folder as a sibling to the `src` folder so that
it doesn't get included in the driver package, so care should be
taken when setting up the LUA PATH to run these tests.