Bugfix: Correctly identify migrated lights
This is a regression that showed up during the refactor that
is causing migrated lights to be non-functional with these changes,
and it is the source of the community issues for which we've had
to hold the driver on beta for a few releases.

This bug is covered by one of the new Busted specs.