Include embedded config for profiles using temperatureAlarm
The temperatureAlarm capability supports four enum values: freeze,
cleared, heat, and rateOfRise. Currently, all driver profiles that
support this capability exclude rateOfRise since it is unused and
will come up in automations if not excluded, and it is the same case
for these dishwasher profiles. Since this enum is not used in the
dishwasher subdriver, it should be excluded just as is done in other
driver profiles.