chore: First integration tests for Hue
The next Lua Libs API release will have rudimentary support for testing
LAN type edge drivers, so we're starting to write some integration tests
for Philips Hue in advance of this.

The current state of the in-development Lua Libs supports basic TCP client/server
REST-ful interactions, so we've started writing some tests to exercise that.