Enable native handlers for drivers that can use them
only switch and switchLevel capabilities will be supported for Matter,
Zigbee, and Z-Wave devices in the 0.54.x firmware. Using this
functionality give a huge latency improvement for handling these
capability commands which is important when a driver is handling several
commands all at once (i.e. a routine is triggered that toggles a bunch
of switches)