Matter Bridge Aqara Cube Support with Custom Capability
Currently, the Generic EdgeDriver applied to the Matter Bridge Aqara Cube is implemented by mapping six faces into individual components,
making it difficult to check the event occurrence of individual faces on one screen.

So this commit wants to improve the problem by mapping the event from Matter Bridge with Aqara Cube Custom Capability developed by Zigbee.
This means that there is a limitation that the Action Event of the Cube that the Matter Bridge does not give can not be processed.

 - REQ-15926, REQ-16285, IOTE-4217, IOTE-4266

Signed-off-by: donghoon-ryu <donghoon.ryu@samsung.com>