Matter Switch: add child profile overrides
Add the ability to override the child profile that would be assigned
based on the child's device type. This is done because some devices
have the "On/Off Plug" device type, but it would make more sense
for them to appear with the switch icon as indicated in their
WWST fingerprint.