Matter-Switch: Improve color temp conversion rounding
This change adjusts the way that the color temperature bounds are
rounded while converting them from mireds into Kelvin. Previously, they
were rounded to the nearest hundreds value and this sometimes resulted
in the range displayed in the ST app being larger than the range
supported by the device. This change will prevent this from occurring by
shortening the range, if needed, to be within the range supported by the
device.