Prevent multiple reads of MinSetpointDeadBand (#1686)
The driver currently attempts to only read MinSetpointDeadBand once
during the first time device_init is called, but the
MIN_SETPOINT_DEADBAND_CHECKED field is not persisted, so the read
request is sent during each device init. This is contributing to a spike
in InteractionQueueFull exceptions. This change ensures that the field
is persisted so that the read request is only sent once.