fix: Don't crash on group update on nil data
It seems that there are situations where a group update will contain
data that does not map to cached data in the driver. This propagates
as passing `nil` values to capability constructors, which in turn results
in constraint violation errors during the pre-emit checks in the Lua Libs.

This PR does two things:

1. It adds additional logging around this path so we can gather information, as
I cannot reproduce this on my setup.

2. It skips the event emit if any of the data isn't valid.