fix: luacov-cobertura dependency acquisition
The version of `luacov-cobertura` on Luarocks is no longer available
on `github`, and the rockspec we're downloading uses the GitHub source URL
in its source definition, so we're failing to acquire it.

There's an alternative rockspec for this package, that's even recommended by the
README of `luacov`, but it's not on Luarocks. So we acquire it from its release
tag on GitHub.

The `sed` one-liners are because the rockspec from the tag reference the artifact
from `master` instead of the artifact from the tag itself, and the version number
wasn't properly incremented.