fix: Remove incorrect dimming logic.
This was some of the first code I wrote for the Hue driver, and I had
a misunderstanding of the `min_dimming` property returned by the API.

The `min_dimming` value in the API is informative, and represents a percent
of the total lumens the device is capable of outputting that you will get
if you request the lowest possible brightness for the bulb (1%).

We were interpreting it as a lower bound on the values to send, which is
not correct.

This PR removes the lower bound checks, and instead clamps all dimming commands
to the range of [1,100] inclusive.